Rename package data to repository
