+++ taco-cloud/src/main/java/by/kiselevich/tacocloud/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package by.kiselevich.tacocloud.data;
+package by.kiselevich.tacocloud.repository;
 
 import by.kiselevich.tacocloud.model.Order;
 import by.kiselevich.tacocloud.model.User;
