+++ taco-cloud/src/main/java/by/kiselevich/tacocloud/repository/TacoRepository.java
@@ -1,4 +1,4 @@
-package by.kiselevich.tacocloud.data;
+package by.kiselevich.tacocloud.repository;
 
 import by.kiselevich.tacocloud.model.Taco;
 import org.springframework.data.repository.CrudRepository;
